refactor: split vmlinux.h into arch-specific files
Replace the single vmlinux.h with vmlinux_x86.h and vmlinux_arm64.h.
The C source conditionally includes the correct one based on the
__TARGET_ARCH_* define. This prevents cross-arch contamination where
an ARM64-generated vmlinux.h was used for x86_64 compilation.

Both files are committed and tagged as generated. To regenerate:
  bpftool btf dump file /sys/kernel/btf/vmlinux format c > pkg/ebpf/bpf/vmlinux_<arch>.h
