+++ .gitignore
@@ -30,4 +30,6 @@ kubeconfig
 *.kubeconfig
 ./kloak
 
-./bin
+# Generated kernel headers (the arch-specific vmlinux_*.h files ARE committed)
+pkg/ebpf/vmlinux.h
+pkg/ebpf/bpf/vmlinux.h
